Changes to move files to config and email folder
